backend/requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
pydantic==2.12.5
pydantic-settings==2.9.1
httpx==0.28.1
python-multipart==0.0.20
python-Levenshtein==0.27.3
langgraph==1.1.3
langchain-core==1.2.20
langgraph-checkpoint==4.0.1
langgraph-sdk==0.3.12
pytest==8.4.1
pytest-asyncio==1.0.0
sqlalchemy==2.0.48
pymssql==2.3.13
alembic==1.18.4